Extract shared LiveStore adapter for use by in-browser agents
- Create shared adapter instance
- Handle reset persistence logic at module level
